pyproject.toml
[project]
name = "jaguar-reid"
version = "0.1.0"
description = ""
authors = [
    {name = "Vincent Eichhorn",email = "vincent.eichhorn@student.hpi.uni-potsdam.de"},
    {name = "Josef Pribbernow",email = "josef.pribbernow@student.hpi.uni-potsdam.de"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "torch (>=2.9.1,<3.0.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "timm (>=1.0.24,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "wandb (>=0.24.0,<0.25.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psutil (>=7.2.1,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.1.0,<8.0.0)"
]

[tool.poetry]
packages = [{include = "jaguar"}]
